% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqc.R
\name{best_model}
\alias{best_model}
\title{select best model given a list of models and a base model}
\usage{
best_model(model_list, fm_base)
}
\examples{
ratio_rho_list <-  multi_ratio_rho(model_list, fm_base)
best_model(ratio_rho_list, model_list)
}

