<<<<<<< HEAD
% Generated by roxygen2 (4.1.1): do not edit by hand
=======
% Generated by roxygen2: do not edit by hand
>>>>>>> master
% Please edit documentation in R/detectFileFormat.R
\name{detectFileFormat}
\alias{detectFileFormat}
\title{Detect file format}
\usage{
detectFileFormat(files)
}
\arguments{
\item{files}{charactor vector of file names}
}
\value{
list of FastqFile and BamFiles objects
}
\description{
Detect file format
}
\examples{
folder <- system.file(package="ShortRead", "extdata/E-MTAB-1147")
files <- list.files(full.names=TRUE, path=folder)
input <- detectFileFormat(files)
sapply(input, class)
}

