% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqc.R
\name{make_model_list}
\alias{make_model_list}
\title{Make a list of models staring from a formula and a formula_base}
\usage{
make_model_list(formula, fm_base, data, mc.cores, ...)
}
\examples{
make_model_list (formula =  stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
            formula_base = stack.loss~1,
            data = stackloss, tau = .5, method = "fn")
}

