% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqc.R
\name{multi_ratio_rho}
\alias{multi_ratio_rho}
\title{compute ratio_rho for a list of models compared to a base model}
\usage{
multi_ratio_rho(model_list, fm_base)
}
\examples{
# List of models
model_list <- list(
   fm1 = rq(formula = stack.loss~Air.Flow, data = stackloss),
   fm2 = rq(formula = stack.loss~Water.Temp, data = stackloss),
   fm3 = rq(formula = stack.loss~Acid.Conc., data = stackloss))
# Base model

#comparison
multi_ratio_rho(model_list, fm_base)
}

