% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqc.R
\name{rq_try}
\alias{rq_try}
\title{try rq model and if error returns NULL}
\usage{
rq_try(..., silent = TRUE)
}
\examples{
df <- data.frame(x = rep(1, 100), y = rnorm(100))
fm <- try(rq_try(y~x , data = df))
class(fm)
}

