% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqmss.R
\name{rqci}
\alias{rqci}
\title{Compare all quantile regression (rq) with i variables}
\usage{
rqci(formula, formula_base, data, tau, method, ...)
}
\arguments{
\item{formula}{Formula for the full model}

\item{formula_base}{Formula for teh base model}

\item{data}{Model data frame}

\item{tau}{a quantile for quantile regression}

\item{method}{Model method for quantile regression}

\item{...}{additional parameters for function `qr`}
}
\value{
data frame of five colums:
* formula character: tested formula
* formula_base character: base formula
* score double: model perfornace indicator
* formula_best integer: 0/1
}
\description{
compute all models starting from formula base and adding all variables one by one
Return a between models comparison statistics.
}
\examples{
rqci(
 formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
 formula_base = stack.loss ~1,
 data = stackloss,
 tau = .5, method = 'fn')
## i = 2
rqci(
 formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
 formula_base = stack.loss ~Air.Flow,
 data = stackloss,
 tau = .5, method = 'fn')
## i = 3
rqci(
 formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
 formula_base = stack.loss~Water.Temp,
 data = stackloss,
 tau = .5, method = 'fn')
 @export
}

