% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqmss.R
\name{rqcn}
\alias{rqcn}
\title{Compare all quantile regression (rq) with i=1:n variables}
\usage{
rqcn(formula, formula_base, data, tau, method, n_vars = NULL, n_sample = -1)
}
\arguments{
\item{formula}{Formula for the full model}

\item{formula_base}{Formula for the base model}

\item{data}{Model data frame}

\item{tau}{a quantile for quantile regression}

\item{method}{Model method for quantile regression}
}
\value{
formula object corrensponding to the best model
}
\description{
Starting from formula base  with i terms  adds variables one by one, computes models
and returns the formulae for all best models with number of variables ranging
between i and n with n being the number of variables in  formula
}
\examples{
# Example 1
rqcn( formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
formula_base = stack.loss ~1,
data = stackloss,
tau = .5, method = 'fn')
# Example 2
rqcn( formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
formula_base = stack.loss ~1,
n_vars = 2,
data = stackloss,
tau = .5, method = 'fn')
}

