% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rqcQA-methods.R
\docType{methods}
\name{rqcQA}
\alias{rqcQA}
\alias{rqcQA,BamFile-method}
\alias{rqcQA,FastqFile-method}
\alias{rqcQA,character-method}
\alias{rqcQA,list-method}
\title{Quality Assessment Rqc function}
\usage{
rqcQA(x, sample = TRUE, n = 1e+06, group = rep("None", length(x)),
  top = 10, ...)

\S4method{rqcQA}{list}(x, sample, n, group, top, workers = multicoreWorkers())

\S4method{rqcQA}{character}(x, sample = TRUE, n = 1e+06,
  group = rep("None", length(x)), top = 10, workers = multicoreWorkers())

\S4method{rqcQA}{BamFile}(x, sample, n, group, top)

\S4method{rqcQA}{FastqFile}(x, sample, n, group, top)
}
\arguments{
\item{x}{input file(s)}

\item{sample}{It reads a random sample from files if this parameter is TRUE.}

\item{n}{Number of sequences to read from each input file.  This represents
sample size if 'sample' parameter is TRUE, if not represents the chunk size
to read on each iteration.  Default is read a sample of one million
sequences from each input file.}

\item{group}{group name for each input file.}

\item{top}{number of top over-represented reads. Default is 10 reads.}

\item{...}{other parameters}

\item{workers}{number of parallel workers}
}
\value{
A named list of \code{RqcResultSet} objects, each one represents a
file.
}
\description{
Process a set of files and returns a list of quality control data.  Files
must be FASTQ format, compressed or not.
}
\details{
Input files are read using \code{FastStreamer} and \code{FastSampler}
classes of \code{\link{ShortRead}} package.  Process multiple files in
parallel using \code{\link{bplapply}} function of \code{\link{BiocParallel}}
package.
}
\section{Methods (by class)}{
\itemize{
\item \code{list}: process a list of \code{\link[ShortRead]{FastqFile}} and
\code{\link[Rsamtools]{BamFile}} objects.

\item \code{character}: automatically detects file format
(using \code{\link{detectFileFormat}} function) of input files then process.

\item \code{BamFile}: process only one BAM file.

\item \code{FastqFile}: process only one FASTQ file.
}}
\examples{
checkpoint("Rqc", path=system.file(package="Rqc", "extdata"), {
  folder <- system.file(package="ShortRead", "extdata/E-MTAB-1147")
  files <- list.files(full.names=TRUE, path=folder)
  rqcResultSet <- rqcQA(files, workers=1)
}, keep="rqcResultSet")
rqcReadQualityPlot(rqcResultSet)
}
\author{
Welliton Souza
}
\seealso{
\code{\link{rqc}}
}

