\name{RqcResultSet-class}
\docType{class}
\alias{RqcResultSet-class}
\alias{RqcResultSet}

\title{ Quality control data }
\description{
  \code{RqcResultSet} class represents quality assessment data created by
  \code{rqc} and \code{rqcQA} functions. This class extends \code{.QA} class
  from \code{ShortRead} package.
}

\seealso{
\code{\link{perReadQuality}} returns read mean quality frequency table.

\code{\link{perReadWidth}} returns read width frequency table.

\code{\link{perCycleQuality}} returns cycle-specific quality frequency table.

\code{\link{perCycleBasecall}} returns cycle-specific base call frequency table.

\code{\link{perFileInformation}} returns informations of the files.
}
