% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqcShinyReport.R
\name{rqcShinyReport}
\alias{rqcShinyReport}
\title{Interactive Quality Control Report}
\usage{
rqcShinyReport(rqcResultSet)
}
\arguments{
\item{rqcResultSet}{list of \code{\link{RqcResultSet-class}} objects}
}
\value{
function
}
\description{
This function runs a Shiny web application of interactive Rqc report. This is
useful for large amount of files and sample groups.
}
\examples{
checkpoint("Rqc", path=system.file(package="Rqc", "extdata"), {
  folder <- system.file(package="ShortRead", "extdata/E-MTAB-1147")
  files <- list.files(full.names=TRUE, path=folder)
  rqcResultSet <- rqcQA(files, pair=c(1,1), workers=1)
}, keep="rqcResultSet")
# rqcShinyReport(rqcResultSet)
}
\author{
Welliton Souza
}
