% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detectFileFormat.R
\name{detectFileFormat}
\alias{detectFileFormat}
\title{Detect file format}
\usage{
detectFileFormat(file)
}
\arguments{
\item{file}{file name}
}
\value{
FastqFile or BamFiles objects
}
\description{
Detect file format
}
\examples{
folder <- system.file(package="ShortRead", "extdata/E-MTAB-1147")
files <- list.files(full.names=TRUE, path=folder)
input <- lapply(files, detectFileFormat)
sapply(input, class)
}

