% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqmss.R
\name{rqc2}
\alias{rqc2}
\title{Compare two quantile regression (rq)  models}
\usage{
rqc2(formula, formula_base, data, tau, method)
}
\arguments{
\item{formula}{Formula for the full model}

\item{formula_base}{Formula for teh base model}

\item{data}{Model data frame}

\item{method}{Model method for quantile regression}

\item{Model}{quantile for quantile regression.}
}
\value{
double: model performance indicator
}
\description{
compute model according to  formula and model_base according to formula_base.
Return a between models comparison statistics.
}
\examples{
rqc2(
 formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.+noise,
 formula_base = stack.loss ~1,
 data = stackloss,
 tau = .5, method = 'fn')
}

