% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rqmss.R
\name{rqci}
\alias{rqci}
\title{Compare all quantile regression (rq) with i variables}
\usage{
rqci(i, formula, formula_base, data, tau, method)
}
\arguments{
\item{formula}{Formula for the full model}

\item{formula_base}{Formula for teh base model}

\item{data}{Model data frame}

\item{tau}{a quantile for quantile regression}

\item{method}{Model method for quantile regression}
}
\value{
formula object corrensponding to the best model
}
\description{
compute all models starting from formula base and adding all variables one by one
Return a between models comparison statistics.
}
\examples{
 rqci( i = 2, formula = stack.loss~Air.Flow+Water.Temp+ Acid.Conc.,
     formula_base = stack.loss ~1,
     data = stackloss,
     tau = .5, method = 'fn')
}

